Rely on autoboxing for entity count stack.

+++ metafix/src/main/java/org/metafacture/metafix/Metafix.java
@@ -239,7 +239,7 @@ public class Metafix implements StreamPipe<StreamReceiver>, Maps {
         flattener.startRecord(identifier);
         entityCountStack.clear();
         entityCount = 0;
-        entityCountStack.add(Integer.valueOf(entityCount));
+        entityCountStack.add(entityCount);
         recordIdentifier = identifier;
         entities = new ArrayList<>();
     }
@@ -317,13 +317,13 @@ public class Metafix implements StreamPipe<StreamReceiver>, Maps {
         addValue(name, value);
         entities.add(value);
 
-        entityCountStack.push(Integer.valueOf(++entityCount));
+        entityCountStack.push(++entityCount);
         flattener.startEntity(name);
     }
 
     @Override
     public void endEntity() {
-        entityCountStack.pop().intValue();
+        entityCountStack.pop();
         flattener.endEntity();
     }
